requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
Django==5.2.7
django-auth-ldap==5.2.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyspnego==0.12.0
python-ldap==3.4.5
pywinrm==0.5.0
requests==2.32.5
requests_ntlm==1.3.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
xmltodict==1.0.2
